update mass int -> double

+++ src/main/java/com/fpt/sep490/dto/ActiveDateRequestDTO.java
@@ -1,4 +1,4 @@
 package com.fpt.sep490.dto;
 
-public record ActiveDateRequestDTO(long employeeId, String dayActive, int mass, String note) {
+public record ActiveDateRequestDTO(long employeeId, String dayActive, double mass, String note) {
 }
